+++ BW/src/bw/body/io/clsSensorActuatorBaseExt.java
@@ -24,10 +24,10 @@ public abstract class clsSensorActuatorBaseExt extends clsSensorActuatorBase{
 		moBaseIO=poBaseIO;
 	}
 
-	protected void registerEnergyConsumption(float prValue) {
+	protected void registerEnergyConsumption(double prValue) {
 		moBaseIO.registerEnergyConsumption(getUniqueId(), prValue);
 	}
-	protected void registerEnergyConsumptionOnce(float prValue) {
+	protected void registerEnergyConsumptionOnce(double prValue) {
 		moBaseIO.registerEnergyConsumptionOnce(getUniqueId(), prValue);
 	}
